+++ app/scripts/controllers/transactions/lib/recipient-blacklist-checker.js
@@ -11,12 +11,7 @@ export default {
   @param {string} account
 */
 function checkAccount (networkId, account) {
-  const MAINNET_LANCHED =
-    new Date().getTime() >
-    new Date(
-      'Thu Oct 29 2020 00:10:00 GMT+0800 (China Standard Time)'
-    ).getTime()
-  const mainnetId = MAINNET_LANCHED ? 1029 : 2
+  const mainnetId = 1029
   if (networkId !== mainnetId) {
     return
   }