+++ ui/app/pages/routes/index.js
@@ -375,15 +375,7 @@ class Routes extends Component {
     let name
 
     if (providerName === 'mainnet') {
-      name = this.context.t('connectingToMainnet').replace('Oceanus', 'Tethys')
-      const MAINNET_LANCHED =
-        new Date().getTime() >
-        new Date(
-          'Thu Oct 29 2020 00:10:00 GMT+0800 (China Standard Time)'
-        ).getTime()
-      if (MAINNET_LANCHED) {
-        name = name.replace('Oceanus', 'Tethys')
-      }
+      name = this.context.t('connectingToMainnet')
     } else if (providerName === 'testnet') {
       name = this.context.t('connectingToTestnet')
     } else {
@@ -400,15 +392,7 @@ class Routes extends Component {
     let name
 
     if (providerName === 'mainnet') {
-      const MAINNET_LANCHED =
-        new Date().getTime() >
-        new Date(
-          'Thu Oct 29 2020 00:10:00 GMT+0800 (China Standard Time)'
-        ).getTime()
       name = this.context.t('mainnet')
-      if (MAINNET_LANCHED) {
-        name = name.replace('Oceanus', 'Tethys')
-      }
     } else if (providerName === 'testnet') {
       name = this.context.t('testnet')
     } else if (providerName === 'localhost') {